Move the process_frame logic to rav1e.rs.
Part of #938.
